feat: leaderboard prettifyHeaders
