+++ packages/ui-kit/src/components/GroupBy/GroupBy.tsx
@@ -6,7 +6,8 @@ import { Checkbox } from '@radix-ui/themes'
 import { Option, OptionValue, Select } from '../Select'
 
 import componentStyles from './GroupBy.module.scss'
-import { buildFormatter, getGroupByLabel, prettifyName } from './utils'
+import { buildFormatter, getGroupByLabel } from './utils'
+import { prettifyName } from '../shared.utils'
 import { Divider } from '../Divider'
 import { Loader } from '../Loader'
 import { GroupByProps } from './GroupBy.types'
