+++ packages/ui-kit/src/components/GroupBy/utils.ts
@@ -10,37 +10,6 @@ export function getGroupByLabel(selectedColumns: string[]) {
   return selectedColumns.length
 }
 
-/**
- * It receives a camelCase or snake_case column name and returns a readable format
- * @example
- * prettifyName('camelCase') => 'Camel Case'
- * prettifyName('snake_case') => 'Snake Case'
- * @param name - Column name
- * @returns Prettified column name
- */
-export function prettifyName(name: string | number) {
-  // If the name is a number, return it as a string
-  if (typeof name === 'number') {
-    return name.toString()
-  }
-
-  // If the name is all uppercase, return it as is
-  if (name === name.toUpperCase()) {
-    return name
-  }
-
-  const camelCaseFormatted = name.replace(/([A-Z])/g, ' $1').trim()
-
-  const snakeCaseFormatted = camelCaseFormatted.replace(/[_-]/g, ' ').trim()
-
-  const capitalizedWords = snakeCaseFormatted
-    .split(' ')
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
-    .join(' ')
-
-  return capitalizedWords
-}
-
 export function buildFormatter(nameFormatter?: (name: string) => string) {
   return function (name: string | number) {
     if (typeof name === 'number') {
